+++ Source/RethinkDb.Driver/Ast/Util.cs
@@ -39,38 +39,10 @@ namespace RethinkDb.Driver.Ast
                 return ast;
             }
 
-            //GitHub Issue #21 - Request - Allow Inserting JObject
             var token = val as JToken;
             if( token != null )
             {
-                //First way to do it is to convert the whole thing into JSON and have
-                //the server parse it. One thing we'd need to watch out for is 
-                //the DateTime conversion. ReqlDateTimeConverter would need to change
-                //so that it directly serializes to $reql_type$:TIME instead of an AST
-                //Iso8601(string) with [49,...]. It's probably more proper to serialize 
-                //to $reql_type$ than a ReQL Iso8601(string)[49, ...].
-                //var json = token.ToString(Formatting.None, Converter.Serializer.Converters.ToArray());
-                //return new Json(json);
                 return new Poco(token);
-
-                //Another way to do it: De-construct the JObject like we do to an IDictionary...
-                //Pro: More complete
-                //Con: Easy to get wrong, lots more code, code-duplication with IDictionary
-                //     
-                //if (token.Type == JTokenType.Object)
-                //{
-                //    var jobj = val as IDictionary<string, JToken>;
-                //    var obj = new Dictionary<string, ReqlAst>();
-                //    foreach (var t in jobj)
-                //    {
-                //        obj[t.Key] = ToReqlAst(t.Value);
-                //    }
-                //    return MakeObj.fromMap(obj);
-                //}
-                //else if (token.Type == JTokenType.Bytes) .. and for each supported JSON native type.
-                //{ 
-                //
-                //}
             }
 
             var lst = val as IList;
@@ -107,20 +79,16 @@ namespace RethinkDb.Driver.Ast
                 return Func.FromLambda(del);
             }
 
-
-            if( val is DateTime )
+            var dt = val as DateTime?;
+            if (dt != null)
             {
-                var dt = (DateTime)val;
-                var isoStr = dt.ToString("o");
-                return Iso8601.FromString(isoStr);
+                return new Poco(dt);
             }
-            if( val is DateTimeOffset )
+            var dto = val as DateTimeOffset?;
+            if (dto != null)
             {
-                var dt = (DateTimeOffset)val;
-                var isoStr = dt.ToString("o");
-                return Iso8601.FromString(isoStr);
+                return new Poco(dto);
             }
-
 
             var @int = val as int?;
             if( @int != null )